requirements.txt
# requirements.txt
Django>=4.0,<5.0
requests>=2.20.0
python-dotenv>=0.19.0
gunicorn
dj-database-url
# psycopg2-binary # If using PostgreSQL

django-jazzmin>=2.6,<3.0
whitenoise[brotli]>=6.0.0 # Add this line